first 98 fib nos

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,14 +10,14 @@
 
 int main(void)
 {
-	unsigned long long int fib1 = 1, fib2 = 1, fib_sum;
+	unsigned long int fib1 = 1, fib2 = 1, fib_sum;
 	int i = 0;
 
 	while (i < 98)
 	{
 		fib_sum = fib1 + fib2;
 
-		printf("%llu", fib2);
+		printf("%lu", fib2);
 		if (i != 97)
 		{
 			printf(", ");
